Added copyright to everything, via the clever IDEA functionality.
This system uses the new exception mapper functionality. Lowlevel have been made to work with this, the rest compiles but lack some the mapper implementations.

Moved the wsdl files from the implementation to the interface. This works for all builds now.

Renamed build-includes.xml to build-common.xml to keep it in line with the project names. Perhaps the build-includes.xml project name should change instead...

Introduced a build-dependencies.xml that handles much of the dependency work for ant. All the big ant includes not depends on this.